Got the code reviewed. Fixed comments.

+++ map.h
@@ -17,30 +17,32 @@ extern "C" {
     map_t* map_t_alloc        (int (*p_comparator)(void*, void*));
 
     /***************************************************************************
-    * If p_map contains the key p_key, associates it with value p_value and    *
-    * returns the old value of that key.                                       *
+    * If the map does not contain the key, inserts it in the map and associates*
+    * the value with it, returning NULL. Otherwise, the value is updated and   *
+    * the old value is returned.                                               *  
     ***************************************************************************/ 
     void*  map_t_put          (map_t* p_map, void* p_key, void* p_value);
 
     /***************************************************************************
-    * Returns a positive value if p_key is mapped to some value in this map.   *
+    * Returns true if the key is mapped in the map.                            *
     ***************************************************************************/
     bool   map_t_contains_key (map_t* p_map, void* p_key);
 
     /***************************************************************************
-    * Returns the value associated with the p_key, or NULL if p_key is not     *
+    * Returns the value associated with the key, or NULL if the key is not     *
     * mapped in the map.                                                       *
     ***************************************************************************/
     void*  map_t_get          (map_t* p_map, void* p_key);
 
     /***************************************************************************
-    * If p_key is mapped in the map, removes the mapping and returns the value *
-    * of that mapping. If the map did not contain the mapping, return NULL.    *
+    * If the key is mapped in the map, removes the mapping and returns the value*
+    * of that mapping. If the map did not contain the mapping, returns NULL.   *
     ***************************************************************************/ 
     void*  map_t_remove       (map_t* p_map, void* p_key);
 
     /***************************************************************************
-    * Removes all the contents of the map.                                     * 
+    * Removes all the contents of the map. Deallocates the map structures. The *
+    * client user is responsible for deallocating the actual contents.         *  
     ***************************************************************************/ 
     void   map_t_clear        (map_t* p_map);
 
@@ -48,16 +50,17 @@ extern "C" {
     * Returns the size of the map, or namely, the amount of key/value mappings *
     * in the map.                                                              *
     ***************************************************************************/ 
-    int    map_t_size         (map_t* p_map);
+    size_t map_t_size         (map_t* p_map);
 
     /***************************************************************************
-    * Checks that the map maintains the AVL-tree invariant.                    *
+    * Checks that the map maintains the AVL-tree property.                     *
     ***************************************************************************/  
     bool   map_t_is_healthy   (map_t* p_map);
 
     /***************************************************************************
     * Deallocates the entire map. Only the map and its nodes are deallocated.  *
-    * The user is responsible to deallocate the actual data stored in the map. *
+    * The user is responsible for deallocating the actual data stored in the   * 
+    * map.                                                                     *
     ***************************************************************************/ 
     void   map_t_free         (map_t* p_map);
 
@@ -69,17 +72,17 @@ extern "C" {
     /***************************************************************************
     * Returns the number of keys not yet iterated over.                        *
     ***************************************************************************/ 
-    int             map_iterator_t_has_next     (map_iterator_t* p_iterator);
+    size_t          map_iterator_t_has_next     (map_iterator_t* p_iterator);
 
     /***************************************************************************
-    * Returns the next entry in the iteration order.                           *
+    * Loads the next mapping in the iteration order.                           *
     ***************************************************************************/  
     bool            map_iterator_t_next         (map_iterator_t* p_iterator, 
                                                  void** pp_key, 
                                                  void** pp_value);
 
     /***************************************************************************
-    * Returns a positive integer if the map was modified during the iteration. *
+    * Returns a true if the map was modified during the iteration.             *
     ***************************************************************************/  
     bool            map_iterator_t_is_disturbed (map_iterator_t* p_iterator);
