+++ unordered_map.h
@@ -12,7 +12,8 @@ extern "C" {
     typedef struct unordered_map_iterator_t unordered_map_iterator_t;
 
     /***************************************************************************
-    * Allocates a new, empty map with given comparator function.               *
+    * Allocates a new, empty map with given hash function and given equality   *
+    * testing function.                                                        * 
     ***************************************************************************/ 
     unordered_map_t* unordered_map_t_alloc 
            (size_t initial_capacity,
@@ -21,8 +22,9 @@ extern "C" {
             bool (*p_equals_function)(void*, void*));
 
     /***************************************************************************
-    * If p_map contains the key p_key, associates it with value p_value and    *
-    * returns the old value of that key.                                       *
+    * If p_map does not contain the key p_key, inserts it in the map,          *
+    * associates p_value with it and return NULL. Otherwise updates the value  *
+    * and returns the old value.                                               * 
     ***************************************************************************/ 
     void*  unordered_map_t_put             (unordered_map_t* p_map, 
                                             void* p_key, 
@@ -43,7 +45,7 @@ extern "C" {
 
     /***************************************************************************
     * If p_key is mapped in the map, removes the mapping and returns the value *
-    * of that mapping. If the map did not contain the mapping, return NULL.    *
+    * of that mapping. If the map did not contain the mapping, returns NULL.   *
     ***************************************************************************/ 
     void*  unordered_map_t_remove          (unordered_map_t* p_map, 
                                             void* p_key);
@@ -57,21 +59,23 @@ extern "C" {
     * Returns the size of the map, or namely, the amount of key/value mappings *
     * in the map.                                                              *
     ***************************************************************************/ 
-    int    unordered_map_t_size            (unordered_map_t* p_map);
+    size_t unordered_map_t_size            (unordered_map_t* p_map);
 
     /***************************************************************************
-    * Checks that the map maintains the AVL-tree invariant.                    *
+    * Checks that the map is in valid state.                                   *
     ***************************************************************************/  
     bool   unordered_map_t_is_healthy      (unordered_map_t* p_map);
 
     /***************************************************************************
     * Deallocates the entire map. Only the map and its nodes are deallocated.  *
-    * The user is responsible to deallocate the actual data stored in the map. *
+    * The user is responsible for deallocating the actual data stored in the   *
+    * map.                                                                     *
     ***************************************************************************/ 
     void   unordered_map_t_free            (unordered_map_t* p_map);
 
     /***************************************************************************
-    * Returns the iterator over the map. The entries are iterated in order.    *
+    * Returns the iterator over the map. The entries are iterated in insertion *
+    * order.                                                                   * 
     ***************************************************************************/  
     unordered_map_iterator_t* unordered_map_iterator_t_alloc
                              (unordered_map_t* p_map);
@@ -79,17 +83,18 @@ extern "C" {
     /***************************************************************************
     * Returns the number of keys not yet iterated over.                        *
     ***************************************************************************/ 
-    int unordered_map_iterator_t_has_next(unordered_map_iterator_t* p_iterator);
+    size_t unordered_map_iterator_t_has_next
+          (unordered_map_iterator_t* p_iterator);
 
     /***************************************************************************
-    * Returns the next entry in the iteration order.                           *
+    * Loads the next entry in the iteration order.                             *
     ***************************************************************************/  
     bool unordered_map_iterator_t_next(unordered_map_iterator_t* p_iterator, 
                                        void** pp_key, 
                                        void** pp_value);
 
     /***************************************************************************
-    * Returns a positive integer if the map was modified during the iteration. *
+    * Returns a true if the map was modified during the iteration.             *
     ***************************************************************************/  
     bool unordered_map_iterator_t_is_disturbed
         (unordered_map_iterator_t* p_iterator);
